requirements.txt
torch==2.0.1
torchvision==0.15.2
streamlit==1.28.0
Pillow==10.0.1
numpy==1.24.3
plotly==5.17.0
opencv-python==4.8.1.78
